feat: implement admin waiters panel, add paging for displaying workshifts and filter by waiter option

+++ src/main/java/org/example/api/service/WaiterService.java
@@ -4,12 +4,16 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.example.api.dto.WaiterRequest;
 import org.example.api.dto.WaiterResponse;
+import org.example.api.exception.EmailAlreadyTakenException;
+import org.example.api.exception.ResourceInUseException;
 import org.example.api.exception.ResourceNotFoundException;
 import org.example.api.model.Role;
 import org.example.api.model.User;
 import org.example.api.model.Waiter;
 import org.example.api.repository.UserRepository;
 import org.example.api.repository.WaiterRepository;
+import org.example.api.repository.WorkShiftRepository;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
@@ -20,6 +24,8 @@ import java.util.List;
 public class WaiterService {
     private final WaiterRepository waiterRepository;
     private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
+    private final WorkShiftRepository workShiftRepository;
 
     public List<WaiterResponse> findAll(){
         return waiterRepository.findAll().stream().map(this::mapToResponse).toList();
@@ -27,13 +33,19 @@ public class WaiterService {
 
     @Transactional
     public WaiterResponse create(WaiterRequest waiterRequest){
+
+        if (userRepository.existsByEmail(waiterRequest.email())){
+            throw new EmailAlreadyTakenException("Kelner z takim adresem e-mail już istnieje.");
+        }
+
         User user = User.builder()
                 .email(waiterRequest.email())
-                .passwordHash(waiterRequest.password())
+                .passwordHash(passwordEncoder.encode(waiterRequest.password()))
                 .firstName(waiterRequest.firstName())
                 .lastName(waiterRequest.lastName())
                 .phoneNumber(waiterRequest.phoneNumber())
                 .role(Role.ROLE_WAITER)
+                .isActive(true)
                 .build();
 
         Waiter waiter = Waiter.builder()
@@ -42,8 +54,9 @@ public class WaiterService {
                 .speaksEnglish(waiterRequest.speaksEnglish())
                 .build();
 
-       Waiter saved = waiterRepository.save(waiter);
-       return mapToResponse(saved);
+        userRepository.save(user);
+        Waiter saved = waiterRepository.save(waiter);
+        return mapToResponse(saved);
     }
 
     @Transactional
@@ -51,6 +64,10 @@ public class WaiterService {
 
         Waiter waiter = waiterRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Kelner nie istnieje"));
+
+        if (workShiftRepository.existsWorkShiftByWaiterId(waiter.getId())){
+            throw new ResourceInUseException("Nie można usunąć kelnera, który posiada przypisane zmiany.");
+        }
 
         waiterRepository.delete(waiter);
         userRepository.delete(waiter.getUser());
